Refac

+++ pacman-core/src/main/java/de/amr/games/pacman/model/GameModel.java
@@ -33,7 +33,7 @@ public interface GameModel {
     byte RED_GHOST = 0, PINK_GHOST = 1, CYAN_GHOST = 2, ORANGE_GHOST = 3;
 
     /** Game loop frequency, ticks per second. */
-    double FPS = 60;
+    float FPS = 60;
 
     /** Movement speed in pixel/sec. */
     float PPS_AT_100_PERCENT = 73.9f; //TODO this should be 75 but that doesn't work yet
